更新 Python 版本至 3.12；拆分amr.py文件
