+++ .python-version
@@ -1,1 +1,1 @@
-3.9
+3.12
